refactor(settings): split the 1,265-line workspace settings page
Each section becomes its own component under
components/namespace/settings/ (profile, branding, icon picker,
preferences, members table, invite form, administration links, danger
zone), and the members fetch moves into a useNamespaceMembers hook.
The page is now a 138-line orchestrator that derives permissions and
composes the sections.

Markup, copy, classNames, aria labels and mutation calls are copied
verbatim. Three deliberate deltas, all covered in the PR body:
- the members list is read once per refresh instead of twice (the page
  fired two identical listMembers calls and threw half the response
  away),
- a newly invited member now appears in the table without a reload —
  the old invite path refreshed only the enrichment maps, never the
  rows (new test, RED against the old page),
- cancelling the invite form discards the draft instead of restoring
  it on reopen, because the form now unmounts.

Dead code removed: an unused CopyButton and a members loading skeleton
that was unreachable behind the page-level "Loading…" guard.

Closes #1194
