+++ packages/platform-ui/src/app/(app)/[handle]/settings/page.tsx
@@ -1,275 +1,27 @@
 'use client';
 
-import { useState, useMemo, useEffect, useCallback } from 'react';
-import * as Switch from '@radix-ui/react-switch';
-import { useQueryClient } from '@tanstack/react-query';
-import { queryKeys } from '@/lib/query-keys';
-import { routes } from '@/lib/routes';
-import { useParams, useRouter } from 'next/navigation';
+import { useMemo, useState } from 'react';
+import { useParams } from 'next/navigation';
 import Link from 'next/link';
-import {
-  ArrowLeft,
-  Check,
-  ClipboardCopy,
-  LogOut,
-  MailIcon,
-  RotateCcw,
-  Trash2,
-  Upload,
-  Users,
-} from 'lucide-react';
-import { ApiError, mediforce } from '@/lib/mediforce';
+import { ArrowLeft } from 'lucide-react';
 import { useAuth } from '@/contexts/auth-context';
 import { useNamespace } from '@/hooks/use-namespace';
-import { UserProfileLink } from '@/components/user-profile-link';
-import {
-  useLeaveNamespace,
-  useRemoveMember,
-  useUpdateMemberRole,
-  useUpdateNamespace,
-} from '@/hooks/use-namespace-mutations';
-import { WORKSPACE_ICONS, WORKSPACE_ICON_KEYS, getWorkspaceIcon, WORKSPACE_DEFAULT_KEY } from '@/lib/workspace-icons';
-import type { NamespaceMember } from '@mediforce/platform-core';
-import { fileToLogoDataUrl, LOGO_MAX_EDGE_PX, LogoTooLargeError } from '@/lib/logo-image';
-import { readableForegroundTriple } from '@/lib/brand-color';
-import { WorkspaceAvatar } from '@/components/workspace-avatar';
-import { cn } from '@/lib/utils';
+import { useNamespaceMembers } from '@/hooks/use-namespace-members';
 import { NamespaceSecretsEditor } from '@/components/namespace/namespace-secrets-editor';
-import { WorkspaceDangerDialog, type WorkspaceDangerMode } from '@/components/namespace/workspace-danger-dialog';
-
-type NamespaceMemberWithId = NamespaceMember & { id: string };
-
-type MemberRole = 'member' | 'admin';
-
-interface MemberWithLastSignIn extends NamespaceMemberWithId {
-  email?: string | null;
-  lastSignInTime?: string | null;
-}
-
-interface InviteResult {
-  email: string;
-  emailSent: boolean;
-  isExisting: boolean;
-}
-
-interface ResendResult {
-  email: string;
-  emailSent: boolean;
-}
-
-// Starting values shown in the color pickers when a workspace has no brand
-// color set yet — the hex equivalents of the platform's default `--primary`
-// (172 66% 32%, teal) and `--accent` (38 92% 50%, amber) tokens in globals.css,
-// so "Save colors" without a change persists the current look.
-const DEFAULT_PRIMARY_HEX = '#1c8779';
-const DEFAULT_ACCENT_HEX = '#f59f0a';
-
-function formatDate(isoString: string): string {
-  return new Date(isoString).toLocaleDateString(undefined, {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-  });
-}
-
-function formatLastSignIn(isoString: string): string {
-  return new Date(isoString).toLocaleString(undefined, {
-    year: 'numeric',
-    month: 'short',
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-    timeZoneName: 'short',
-  });
-}
-
-function RoleBadge({ role }: { role: NamespaceMember['role'] }) {
-  const styles: Record<NamespaceMember['role'], string> = {
-    owner: 'bg-amber-500/10 text-amber-600 dark:text-amber-400',
-    admin: 'bg-blue-500/10 text-blue-600 dark:text-blue-400',
-    member: 'bg-muted text-muted-foreground',
-  };
-
-  return (
-    <span
-      className={[
-        'rounded-full px-2 py-0.5 text-[11px] font-medium',
-        styles[role],
-      ].join(' ')}
-    >
-      {role}
-    </span>
-  );
-}
-
-/**
- * Preview chip for a picked brand color. The label colour is derived from the
- * swatch itself, so a light brand colour stays legible instead of showing white
- * text on near-white.
- */
-function BrandSwatch({ color, label }: { color: string; label: string }) {
-  const foreground = readableForegroundTriple(color);
-  return (
-    <span
-      className="inline-flex items-center rounded-md px-3 py-1.5 text-xs font-medium"
-      style={{
-        backgroundColor: color,
-        color: foreground !== null ? `hsl(${foreground})` : undefined,
-      }}
-    >
-      {label}
-    </span>
-  );
-}
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-
-  async function handleCopy() {
-    await navigator.clipboard.writeText(text);
-    setCopied(true);
-    setTimeout(() => setCopied(false), 2000);
-  }
-
-  return (
-    <button
-      type="button"
-      onClick={handleCopy}
-      className="inline-flex items-center gap-1 rounded px-2 py-0.5 text-xs text-muted-foreground hover:text-foreground hover:bg-muted transition-colors"
-      aria-label="Copy to clipboard"
-    >
-      {copied ? <Check className="h-3 w-3" /> : <ClipboardCopy className="h-3 w-3" />}
-      {copied ? 'Copied' : 'Copy'}
-    </button>
-  );
-}
-
-function DefaultWorkspaceSection({ handle }: { handle: string }) {
-  const [isDefault, setIsDefault] = useState(false);
-
-  useEffect(() => {
-    setIsDefault(localStorage.getItem(WORKSPACE_DEFAULT_KEY) === handle);
-  }, [handle]);
-
-  function handleToggle(checked: boolean) {
-    if (checked) {
-      localStorage.setItem(WORKSPACE_DEFAULT_KEY, handle);
-    } else {
-      localStorage.removeItem(WORKSPACE_DEFAULT_KEY);
-    }
-    setIsDefault(checked);
-  }
-
-  return (
-    <div className="mb-10 space-y-4">
-      <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Preferences</h2>
-      <div className="rounded-lg border bg-card px-4 py-5">
-        <div className="flex items-center justify-between gap-4">
-          <div>
-            <p className="text-sm font-medium">Default workspace</p>
-            <p className="text-xs text-muted-foreground mt-0.5">
-              Open this workspace automatically when you sign in.
-            </p>
-          </div>
-          <Switch.Root
-            checked={isDefault}
-            onCheckedChange={handleToggle}
-            className="relative inline-flex h-5 w-9 shrink-0 cursor-pointer items-center rounded-full border-2 border-transparent bg-input transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background data-[state=checked]:bg-primary"
-            aria-label="Set as default workspace"
-          >
-            <Switch.Thumb className="pointer-events-none block h-4 w-4 rounded-full bg-background shadow-lg ring-0 transition-transform data-[state=checked]:translate-x-4 data-[state=unchecked]:translate-x-0" />
-          </Switch.Root>
-        </div>
-      </div>
-    </div>
-  );
-}
+import { DefaultWorkspaceSection } from '@/components/namespace/settings/default-workspace-section';
+import { WorkspaceAdministrationSection } from '@/components/namespace/settings/workspace-administration-section';
+import { WorkspaceDangerZone } from '@/components/namespace/settings/workspace-danger-zone';
+import { WorkspaceMembersSection } from '@/components/namespace/settings/workspace-members-section';
+import { WorkspaceProfileSection } from '@/components/namespace/settings/workspace-profile-section';
 
 export default function WorkspaceConfigPage() {
   const params = useParams();
   const rawHandle = params.handle;
   const handle = Array.isArray(rawHandle) ? rawHandle[0] : (rawHandle ?? '');
-  const router = useRouter();
-  const qc = useQueryClient();
 
   const { user } = useAuth();
   const { namespace, personalHandles, loading: namespaceLoading } = useNamespace(handle);
-
-  // Polled members list via the headless contract — replaces the previous
-  // `onSnapshot(namespaces/{handle}/members)` subscription. Realtime updates
-  // become a small staleness window (next listMembers tick after a mutation);
-  // §"Phase 4 is a swap, not a redesign" accepts this regression.
-  const [realtimeMembers, setRealtimeMembers] = useState<NamespaceMemberWithId[]>([]);
-  const [membersLoading, setMembersLoading] = useState(true);
-
-  const refreshMembers = useCallback(async () => {
-    if (handle === '') return;
-    try {
-      const { members: fetched } = await mediforce.users.listMembers({ namespace: handle });
-      setRealtimeMembers(
-        fetched.map((m) => ({ ...m, displayName: m.displayName ?? undefined, id: m.uid })),
-      );
-    } finally {
-      setMembersLoading(false);
-    }
-  }, [handle]);
-
-  useEffect(() => {
-    void refreshMembers();
-  }, [refreshMembers]);
-
-  // API fetch for lastSignInTime + email + auth-enriched displayName.
-  // Owner member docs created before the createNamespace handler started
-  // persisting displayName lack it locally; the listMembers handler falls
-  // back to the `auth_users` profile name, and we merge that override
-  // here so legacy workspaces show a human name instead of the uid.
-  const [lastSignInMap, setLastSignInMap] = useState<Map<string, string | null>>(new Map());
-  const [emailMap, setEmailMap] = useState<Map<string, string | null>>(new Map());
-  const [displayNameMap, setDisplayNameMap] = useState<Map<string, string | null>>(new Map());
-
-  const fetchLastSignIn = useCallback(async () => {
-    if (handle === '' || user === null) return;
-    try {
-      const { members } = await mediforce.users.listMembers({ namespace: handle });
-      const map = new Map<string, string | null>();
-      const emailMapLocal = new Map<string, string | null>();
-      const displayNameMapLocal = new Map<string, string | null>();
-      for (const member of members) {
-        map.set(member.uid, member.lastSignInTime);
-        emailMapLocal.set(member.uid, member.email);
-        displayNameMapLocal.set(member.uid, member.displayName);
-      }
-      setLastSignInMap(map);
-      setEmailMap(emailMapLocal);
-      setDisplayNameMap(displayNameMapLocal);
-    } catch {
-      // non-fatal — lastSignIn just won't show
-    }
-  }, [handle, user]);
-
-  useEffect(() => {
-    void fetchLastSignIn();
-  }, [fetchLastSignIn]);
-
-  // Merge realtime members with enriched displayName + email + lastSignInTime
-  const members = useMemo((): MemberWithLastSignIn[] => {
-    return realtimeMembers
-      .map((member) => {
-        const enrichedDisplayName = displayNameMap.get(member.uid);
-        return {
-          ...member,
-          displayName: enrichedDisplayName !== undefined && enrichedDisplayName !== null
-            ? enrichedDisplayName
-            : member.displayName,
-          email: emailMap.get(member.uid),
-          lastSignInTime: lastSignInMap.get(member.uid),
-        };
-      })
-      .sort((memberA, memberB) => {
-        const roleOrder: Record<string, number> = { owner: 0, admin: 1, member: 2 };
-        return (roleOrder[memberA.role] ?? 3) - (roleOrder[memberB.role] ?? 3);
-      });
-  }, [realtimeMembers, lastSignInMap, emailMap, displayNameMap]);
+  const { members, loading: membersLoading, refresh: refreshMembers } = useNamespaceMembers(handle);
 
   const currentUserMember = useMemo(
     () =>
@@ -299,258 +51,13 @@ export default function WorkspaceConfigPage() {
   const isPersonal = namespace?.type === 'personal';
   const canDestroyWorkspace = isOwner || isPersonalOwner;
 
-  // ── Workspace profile editing ──────────────────────────────────────────────
-  const [profileDisplayName, setProfileDisplayName] = useState('');
-  const [profileBio, setProfileBio] = useState('');
-  const [profileIcon, setProfileIcon] = useState('Building2');
-  const [profileSaved, setProfileSaved] = useState(false);
-  const updateNamespace = useUpdateNamespace(handle);
-  const savingProfile = updateNamespace.isPending;
-
-  const [primaryColor, setPrimaryColor] = useState(DEFAULT_PRIMARY_HEX);
-  const [accentColor, setAccentColor] = useState(DEFAULT_ACCENT_HEX);
-  const [logoError, setLogoError] = useState<string | null>(null);
-  const [brandSaved, setBrandSaved] = useState(false);
-
-  // Initialise form fields from namespace once loaded
-  useEffect(() => {
-    if (namespace !== null) {
-      setProfileDisplayName(namespace.displayName);
-      setProfileBio(namespace.bio ?? '');
-      setProfileIcon(namespace.icon ?? 'Building2');
-      setPrimaryColor(
-        namespace.brandPrimaryColor !== undefined && namespace.brandPrimaryColor !== ''
-          ? namespace.brandPrimaryColor
-          : DEFAULT_PRIMARY_HEX,
-      );
-      setAccentColor(
-        namespace.brandAccentColor !== undefined && namespace.brandAccentColor !== ''
-          ? namespace.brandAccentColor
-          : DEFAULT_ACCENT_HEX,
-      );
-    }
-  }, [namespace]);
-
-  async function handleLogoChange(event: React.ChangeEvent<HTMLInputElement>) {
-    setLogoError(null);
-    const file = event.target.files?.[0];
-    event.target.value = '';
-    if (file === undefined) return;
-    if (!file.type.startsWith('image/')) {
-      setLogoError('Please choose an image file.');
-      return;
-    }
-    let dataUrl: string;
-    try {
-      dataUrl = await fileToLogoDataUrl(file);
-    } catch (error) {
-      setLogoError(
-        error instanceof LogoTooLargeError ? error.message : 'Image could not be read.',
-      );
-      return;
-    }
-    try {
-      await updateNamespace.mutateAsync({ handle, logo: dataUrl });
-    } catch {
-      setLogoError('Failed to upload logo.');
-    }
-  }
-
-  async function handleRemoveLogo() {
-    setLogoError(null);
-    try {
-      await updateNamespace.mutateAsync({ handle, logo: '' });
-    } catch {
-      setLogoError('Failed to remove logo.');
-    }
-  }
-
-  async function handleSaveBrandColors() {
-    try {
-      await updateNamespace.mutateAsync({ handle, brandPrimaryColor: primaryColor, brandAccentColor: accentColor });
-      setBrandSaved(true);
-      setTimeout(() => setBrandSaved(false), 2500);
-    } catch {
-      // Optimistic snapshot already restored by the hook.
-    }
-  }
-
-  async function handleResetBrandColors() {
-    setPrimaryColor(DEFAULT_PRIMARY_HEX);
-    setAccentColor(DEFAULT_ACCENT_HEX);
-    try {
-      await updateNamespace.mutateAsync({ handle, brandPrimaryColor: '', brandAccentColor: '' });
-    } catch {
-      // Optimistic snapshot already restored by the hook.
-    }
-  }
-
-  async function handleSaveProfile(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    const trimmedName = profileDisplayName.trim();
-    if (trimmedName === '') return;
-    const trimmedBio = profileBio.trim();
-    try {
-      await updateNamespace.mutateAsync({
-        handle,
-        displayName: trimmedName,
-        bio: trimmedBio,
-      });
-      setProfileSaved(true);
-      setTimeout(() => setProfileSaved(false), 2500);
-    } catch {
-      // Optimistic snapshot already restored by the hook; nothing to do.
-    }
-  }
-
-  // ── Icon picker ────────────────────────────────────────────────────────────
-  async function handleIconChange(iconKey: string) {
-    setProfileIcon(iconKey);
-    try {
-      await updateNamespace.mutateAsync({ handle, icon: iconKey });
-    } catch {
-      setProfileIcon(namespace?.icon ?? 'Building2');
-    }
-  }
-
-  // ── Invite form state ──────────────────────────────────────────────────────
-  const [showInviteForm, setShowInviteForm] = useState(false);
-  const [inviteEmail, setInviteEmail] = useState('');
-  const [inviteName, setInviteName] = useState('');
-  const [inviteRole, setInviteRole] = useState<MemberRole>('member');
-  const [inviting, setInviting] = useState(false);
-  const [inviteResult, setInviteResult] = useState<InviteResult | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [resendingUid, setResendingUid] = useState<string | null>(null);
-  const [resendResult, setResendResult] = useState<ResendResult | null>(null);
-  const [resendError, setResendError] = useState<string | null>(null);
-
-  async function handleInvite(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    setError(null);
-
-    const trimmedEmail = inviteEmail.trim().toLowerCase();
-    if (trimmedEmail === '') {
-      setError('Email is required.');
-      return;
-    }
-
-    setInviting(true);
-    try {
-      const data = await mediforce.users.invite({
-        email: trimmedEmail,
-        displayName: inviteName.trim() !== '' ? inviteName.trim() : undefined,
-        namespaceHandle: handle,
-        role: inviteRole,
-        inviterName: user?.name ?? user?.email ?? undefined,
-      });
-
-      setInviteResult({
-        email: data.email,
-        emailSent: data.emailSent,
-        isExisting: data.isExisting,
-      });
-      setShowInviteForm(false);
-      setInviteEmail('');
-      setInviteName('');
-      setInviteRole('member');
-      // Refresh the members list — PR4 swapped the previous onSnapshot
-      // subscription for `useNamespace` react-query, so the cache needs an
-      // explicit invalidation now that the invite mutation isn't a hook.
-      await qc.invalidateQueries({ queryKey: queryKeys.namespace(handle) });
-      void fetchLastSignIn();
-    } catch (err: unknown) {
-      setError(err instanceof Error ? err.message : 'Failed to send invite.');
-    } finally {
-      setInviting(false);
-    }
-  }
-
-  async function handleResendInvite(memberUid: string) {
-    setResendResult(null);
-    setResendError(null);
-    setResendingUid(memberUid);
-    try {
-      const data = await mediforce.users.resendInvite({
-        uid: memberUid,
-        namespaceHandle: handle,
-      });
-      setResendResult({
-        email: data.email,
-        emailSent: data.emailSent,
-      });
-      void fetchLastSignIn();
-    } catch (err: unknown) {
-      // The button is hidden once a member has signed in, but a member who set
-      // a password without ever signing in still isn't pending, so the handler
-      // throws PreconditionFailed. Surface it here — the shared `error` banner
-      // only renders inside the (closed) invite form, so it would be silent.
-      setResendError(
-        err instanceof Error ? err.message : 'Failed to resend invite.',
-      );
-    } finally {
-      setResendingUid(null);
-    }
-  }
-
-  const removeMember = useRemoveMember(handle);
-  const updateMemberRole = useUpdateMemberRole(handle);
-  const leaveNamespace = useLeaveNamespace();
-
   // Surface fail-cases for every workspace-altering action via one banner —
   // member removal, role flips, leave. (Delete and reset report inside their
-  // own dialog.) Optimistic snapshot rollback
-  // hides the underlying error; without this banner an admin would see the
-  // role un-flip and not know why.
+  // own dialog.) Optimistic snapshot rollback hides the underlying error;
+  // without this banner an admin would see the role un-flip and not know why.
   const [dangerError, setDangerError] = useState<string | null>(null);
 
-  async function handleRemoveMember(memberUid: string) {
-    setDangerError(null);
-    try {
-      await removeMember.mutateAsync({ handle, uid: memberUid });
-      void refreshMembers();
-    } catch (err: unknown) {
-      setDangerError(err instanceof Error ? err.message : 'Failed to remove member.');
-    }
-  }
-
-  async function handleToggleRole(memberUid: string, currentRole: string) {
-    setDangerError(null);
-    const nextRole: 'admin' | 'member' = currentRole === 'admin' ? 'member' : 'admin';
-    try {
-      await updateMemberRole.mutateAsync({ handle, uid: memberUid, role: nextRole });
-      void refreshMembers();
-    } catch (err: unknown) {
-      setDangerError(err instanceof Error ? err.message : 'Failed to update member role.');
-    }
-  }
-
-  // ── Danger zone ────────────────────────────────────────────────────────────
-  // A personal workspace cannot be deleted — `getMe` re-bootstraps one for
-  // every user, so the API rejects it (issue #1044). Reset is the destructive
-  // action it does offer.
-  const [dangerDialog, setDangerDialog] = useState<WorkspaceDangerMode | null>(null);
-
-  async function handleLeaveWorkspace() {
-    if (user === null) return;
-    setDangerError(null);
-    try {
-      await leaveNamespace.mutateAsync({ handle });
-      router.push('/workspace-selection');
-    } catch (err: unknown) {
-      if (err instanceof ApiError && err.code === 'precondition_failed') {
-        setDangerError(err.message);
-        return;
-      }
-      setDangerError(err instanceof Error ? err.message : 'Failed to leave workspace.');
-    }
-  }
-
-  const leaving = leaveNamespace.isPending;
-
-  const loading = namespaceLoading || membersLoading;
-
-  if (loading) {
+  if (namespaceLoading || membersLoading) {
     return (
       <div className="flex flex-1 items-center justify-center p-6">
         <div className="text-sm text-muted-foreground animate-pulse">Loading…</div>
@@ -584,549 +91,22 @@ export default function WorkspaceConfigPage() {
           </div>
         </div>
 
-        {/* ── Section 1: Workspace profile ─────────────────────────────────── */}
-        {canEditProfile && (
-          <div className="mb-10 space-y-4">
-            <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Profile</h2>
+        {canEditProfile && <WorkspaceProfileSection handle={handle} namespace={namespace} />}
 
-            {/* Name + bio form */}
-            <div className="rounded-lg border bg-card px-4 py-5">
-              <form onSubmit={handleSaveProfile} className="flex flex-col gap-4">
-                <div className="flex flex-col gap-1.5">
-                  <label htmlFor="workspaceName" className="text-sm font-medium">
-                    Workspace name
-                  </label>
-                  <input
-                    id="workspaceName"
-                    type="text"
-                    value={profileDisplayName}
-                    onChange={(e) => setProfileDisplayName(e.target.value)}
-                    placeholder="My workspace"
-                    className="rounded-md border bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50"
-                    disabled={savingProfile}
-                    required
-                  />
-                </div>
-
-                <div className="flex flex-col gap-1.5">
-                  <label htmlFor="workspaceBio" className="text-sm font-medium">
-                    Description
-                  </label>
-                  <textarea
-                    id="workspaceBio"
-                    value={profileBio}
-                    onChange={(e) => setProfileBio(e.target.value)}
-                    placeholder="What is this workspace used for?"
-                    rows={3}
-                    className="rounded-md border bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50 resize-none"
-                    disabled={savingProfile}
-                  />
-                </div>
-
-                <div className="flex items-center gap-3">
-                  <button
-                    type="submit"
-                    disabled={savingProfile || profileDisplayName.trim() === ''}
-                    className="inline-flex items-center gap-1.5 rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground hover:bg-primary/90 disabled:opacity-50 transition-colors"
-                  >
-                    {savingProfile ? 'Saving…' : 'Save changes'}
-                  </button>
-                  {profileSaved && (
-                    <span className="flex items-center gap-1 text-xs text-green-600 dark:text-green-400">
-                      <Check className="h-3 w-3" />
-                      Saved
-                    </span>
-                  )}
-                </div>
-              </form>
-            </div>
-
-            {/* Branding: logo + brand colors */}
-            {namespace.type === 'organization' && (
-              <div className="rounded-lg border bg-card px-4 py-5 space-y-6">
-                <div>
-                  <h3 className="text-sm font-semibold">Logo</h3>
-                  <p className="text-xs text-muted-foreground mt-0.5">
-                    Upload an image to use instead of the icon. Shown in the sidebar, workspace header, and workspace picker.
-                  </p>
-                  <div className="mt-4 flex items-center gap-6">
-                    <div className="flex h-14 w-14 shrink-0 items-center justify-center overflow-hidden rounded-xl bg-primary/10">
-                      <WorkspaceAvatar
-                        source={namespace.logo}
-                        alt="Workspace logo"
-                        className="h-full w-full object-cover"
-                        fallback={(() => {
-                          const Icon = getWorkspaceIcon(profileIcon);
-                          return <Icon className="h-7 w-7 text-primary" />;
-                        })()}
-                      />
-                    </div>
-                    <div className="flex flex-col gap-2">
-                      <div className="flex items-center gap-2">
-                        <label className="inline-flex cursor-pointer items-center gap-1.5 rounded-md border px-3 py-1.5 text-xs font-medium hover:bg-muted transition-colors">
-                          <Upload className="h-3.5 w-3.5" />
-                          {savingProfile ? 'Uploading…' : 'Upload logo'}
-                          <input
-                            type="file"
-                            accept="image/png,image/jpeg,image/svg+xml,image/webp,image/gif"
-                            className="hidden"
-                            onChange={handleLogoChange}
-                            disabled={savingProfile}
-                          />
-                        </label>
-                        {namespace.logo !== undefined && namespace.logo !== '' && (
-                          <button
-                            type="button"
-                            onClick={handleRemoveLogo}
-                            disabled={savingProfile}
-                            className="inline-flex items-center gap-1.5 rounded-md border border-destructive/30 px-3 py-1.5 text-xs font-medium text-destructive hover:bg-destructive/10 disabled:opacity-50 transition-colors"
-                          >
-                            <Trash2 className="h-3.5 w-3.5" />
-                            Remove
-                          </button>
-                        )}
-                      </div>
-                      <p className="text-[11px] text-muted-foreground">
-                        PNG, JPG, SVG, WebP or GIF · large images are resized to{' '}
-                        {LOGO_MAX_EDGE_PX}px · animated GIFs use their first frame.
-                      </p>
-                      {logoError !== null && <p className="text-[11px] text-destructive">{logoError}</p>}
-                    </div>
-                  </div>
-                </div>
-
-                <div className="border-t" />
-
-                <div>
-                  <h3 className="text-sm font-semibold">Brand colors</h3>
-                  <p className="text-xs text-muted-foreground mt-0.5">
-                    Set the main and auxiliary colors used across the workspace.
-                  </p>
-                  <div className="mt-4 flex flex-wrap items-end gap-6">
-                    <div className="flex flex-col gap-1.5">
-                      <label htmlFor="primaryColor" className="text-xs font-medium">Main color</label>
-                      <div className="flex items-center gap-2">
-                        <input
-                          id="primaryColor"
-                          type="color"
-                          value={primaryColor}
-                          onChange={(e) => setPrimaryColor(e.target.value)}
-                          disabled={savingProfile}
-                          className="h-9 w-12 cursor-pointer rounded border bg-background p-1 disabled:opacity-50"
-                        />
-                        <span className="font-mono text-xs text-muted-foreground uppercase">{primaryColor}</span>
-                      </div>
-                    </div>
-                    <div className="flex flex-col gap-1.5">
-                      <label htmlFor="accentColor" className="text-xs font-medium">Auxiliary color</label>
-                      <div className="flex items-center gap-2">
-                        <input
-                          id="accentColor"
-                          type="color"
-                          value={accentColor}
-                          onChange={(e) => setAccentColor(e.target.value)}
-                          disabled={savingProfile}
-                          className="h-9 w-12 cursor-pointer rounded border bg-background p-1 disabled:opacity-50"
-                        />
-                        <span className="font-mono text-xs text-muted-foreground uppercase">{accentColor}</span>
-                      </div>
-                    </div>
-                    {/* Live preview using the picked colors directly */}
-                    <div className="flex items-center gap-2">
-                      <BrandSwatch color={primaryColor} label="Main" />
-                      <BrandSwatch color={accentColor} label="Auxiliary" />
-                    </div>
-                  </div>
-                  <div className="mt-4 flex items-center gap-3">
-                    <button
-                      type="button"
-                      onClick={handleSaveBrandColors}
-                      disabled={savingProfile}
-                      className="inline-flex items-center gap-1.5 rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground hover:bg-primary/90 disabled:opacity-50 transition-colors"
-                    >
-                      {savingProfile ? 'Saving…' : 'Save colors'}
-                    </button>
-                    <button
-                      type="button"
-                      onClick={handleResetBrandColors}
-                      disabled={savingProfile}
-                      className="rounded-md border px-3 py-2 text-sm font-medium hover:bg-muted disabled:opacity-50 transition-colors"
-                    >
-                      Reset to default
-                    </button>
-                    {brandSaved && (
-                      <span className="flex items-center gap-1 text-xs text-green-600 dark:text-green-400">
-                        <Check className="h-3 w-3" />
-                        Saved
-                      </span>
-                    )}
-                  </div>
-                </div>
-              </div>
-            )}
-
-            {/* Icon picker */}
-            {namespace.type === 'organization' && (
-              <div className="rounded-lg border bg-card px-4 py-5">
-                <h3 className="text-sm font-semibold mb-4">Workspace icon</h3>
-                <div className="flex items-center gap-6">
-                  {/* Preview */}
-                  <div className="flex h-14 w-14 shrink-0 items-center justify-center rounded-xl bg-primary/10">
-                    {(() => {
-                      const Icon = getWorkspaceIcon(profileIcon);
-                      return <Icon className="h-7 w-7 text-primary" />;
-                    })()}
-                  </div>
-                  {/* Grid */}
-                  <div className="grid grid-cols-5 gap-1.5">
-                    {WORKSPACE_ICON_KEYS.map((key) => {
-                      const Icon = WORKSPACE_ICONS[key]!;
-                      return (
-                        <button
-                          key={key}
-                          type="button"
-                          title={key}
-                          onClick={() => handleIconChange(key)}
-                          className={cn(
-                            'flex h-9 w-9 items-center justify-center rounded-lg border-2 transition-colors',
-                            profileIcon === key
-                              ? 'border-primary bg-primary/10 text-primary'
-                              : 'border-transparent bg-muted text-muted-foreground hover:border-border hover:text-foreground',
-                          )}
-                        >
-                          <Icon className="h-4 w-4" />
-                        </button>
-                      );
-                    })}
-                  </div>
-                </div>
-              </div>
-            )}
-          </div>
-        )}
-
-        {/* ── Section 2: Default workspace ─────────────────────────────────── */}
         <DefaultWorkspaceSection handle={handle} />
 
-        {/* ── Section 3: Members ────────────────────────────────────────────── */}
-        <div className="mb-10">
-          <div className="mb-4 flex items-center justify-between gap-4">
-            <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Members</h2>
-            {canManageMembers && !showInviteForm && (
-              <button
-                type="button"
-                onClick={() => { setInviteResult(null); setShowInviteForm(true); }}
-                className="rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground hover:bg-primary/90 transition-colors"
-              >
-                Invite user
-              </button>
-            )}
-          </div>
+        <WorkspaceMembersSection
+          handle={handle}
+          members={members}
+          personalHandles={personalHandles}
+          canManageMembers={canManageMembers}
+          isOwner={isOwner}
+          currentUserId={user?.id}
+          inviterName={user?.name ?? user?.email ?? undefined}
+          onMembersChanged={() => { void refreshMembers(); }}
+          onError={setDangerError}
+        />
 
-          {/* Resend result card */}
-          {resendResult !== null && (
-            <div className="mb-4 rounded-lg border border-blue-200 bg-blue-50 dark:border-blue-800 dark:bg-blue-950/30 px-4 py-4">
-              <div className="flex items-start gap-2">
-                <Check className="h-4 w-4 text-blue-600 dark:text-blue-400 mt-0.5 shrink-0" />
-                <div className="space-y-1 min-w-0">
-                  <p className="text-sm font-medium text-blue-800 dark:text-blue-200">
-                    Invite resent to {resendResult.email}
-                  </p>
-                  <p className="text-sm text-blue-700 dark:text-blue-300">
-                    They can sign in with Google or set a password.
-                  </p>
-                  <p className="text-xs text-blue-600 dark:text-blue-400">
-                    {resendResult.emailSent
-                      ? 'Email sent ✓'
-                      : 'Email not sent — let them know to sign in'}
-                  </p>
-                </div>
-                <button
-                  type="button"
-                  onClick={() => setResendResult(null)}
-                  className="ml-auto text-blue-400 hover:text-blue-600 transition-colors text-lg leading-none"
-                  aria-label="Dismiss"
-                >
-                  ×
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* Resend error card */}
-          {resendError !== null && (
-            <div className="mb-4 rounded-lg border border-destructive/30 bg-destructive/10 px-4 py-4">
-              <div className="flex items-start gap-2">
-                <p className="text-sm font-medium text-destructive">{resendError}</p>
-                <button
-                  type="button"
-                  onClick={() => setResendError(null)}
-                  className="ml-auto text-destructive/60 hover:text-destructive transition-colors text-lg leading-none"
-                  aria-label="Dismiss"
-                >
-                  ×
-                </button>
-              </div>
-            </div>
-          )}
-
-          {/* Invite result card */}
-          {inviteResult !== null && (
-            <div className="mb-4 rounded-lg border border-green-200 bg-green-50 dark:border-green-800 dark:bg-green-950/30 px-4 py-4">
-              <div className="flex items-start gap-2">
-                <Check className="h-4 w-4 text-green-600 dark:text-green-400 mt-0.5 shrink-0" />
-                <div className="space-y-1 min-w-0">
-                  <p className="text-sm font-medium text-green-800 dark:text-green-200">
-                    {inviteResult.isExisting
-                      ? `${inviteResult.email} added to workspace`
-                      : `Invited ${inviteResult.email}`}
-                  </p>
-                  {!inviteResult.isExisting && (
-                    <p className="text-sm text-green-700 dark:text-green-300">
-                      They can sign in with Google or set a password.
-                    </p>
-                  )}
-                  <p className="text-xs text-green-600 dark:text-green-400">
-                    {inviteResult.emailSent
-                      ? 'Email sent ✓'
-                      : inviteResult.isExisting
-                        ? 'Email not sent'
-                        : 'Email not sent — let them know to sign in'}
-                  </p>
-                </div>
-              </div>
-            </div>
-          )}
-
-          {/* Members table */}
-          {membersLoading ? (
-            <div className="space-y-3">
-              {[1, 2, 3].map((index) => (
-                <div
-                  key={index}
-                  className="rounded-lg border bg-card px-4 py-4 animate-pulse flex gap-3"
-                >
-                  <div className="h-8 w-8 rounded-full bg-muted shrink-0" />
-                  <div className="space-y-2 flex-1">
-                    <div className="h-4 w-36 rounded bg-muted" />
-                    <div className="h-3 w-52 rounded bg-muted" />
-                  </div>
-                </div>
-              ))}
-            </div>
-          ) : members.length === 0 ? (
-            <div className="flex flex-col items-center justify-center gap-3 py-16 text-center">
-              <div className="flex h-12 w-12 items-center justify-center rounded-full bg-muted">
-                <Users className="h-6 w-6 text-muted-foreground" />
-              </div>
-              <p className="text-sm text-muted-foreground">No members yet.</p>
-            </div>
-          ) : (
-            <div className="overflow-x-auto rounded-lg border bg-card">
-              <table className="w-full text-sm border-collapse">
-                <thead>
-                  <tr className="border-b bg-muted/50">
-                    <th className="px-4 py-2.5 text-left text-xs font-medium text-muted-foreground whitespace-nowrap">User</th>
-                    {canManageMembers && (
-                      <th className="px-4 py-2.5 text-left text-xs font-medium text-muted-foreground whitespace-nowrap">Email</th>
-                    )}
-                    <th className="px-4 py-2.5 text-left text-xs font-medium text-muted-foreground whitespace-nowrap">Role</th>
-                    <th className="px-4 py-2.5 text-left text-xs font-medium text-muted-foreground whitespace-nowrap">Joined</th>
-                    {canManageMembers && (
-                      <th className="px-4 py-2.5 text-left text-xs font-medium text-muted-foreground whitespace-nowrap">Last sign in</th>
-                    )}
-                    <th className="px-4 py-2.5 text-right text-xs font-medium text-muted-foreground whitespace-nowrap sr-only">Actions</th>
-                  </tr>
-                </thead>
-                <tbody className="divide-y">
-                  {members.map((member) => {
-                    const name = member.displayName ?? member.uid;
-                    const avatar = member.avatarUrl;
-                    const initials = name.includes(' ')
-                      ? `${name.split(' ')[0]?.[0] ?? ''}${name.split(' ')[1]?.[0] ?? ''}`.toUpperCase()
-                      : name.slice(0, 2).toUpperCase();
-
-                    return (
-                      <tr key={member.id} className="hover:bg-muted/30 transition-colors">
-                        {/* User */}
-                        <td className="px-4 py-3 whitespace-nowrap">
-                          <UserProfileLink
-                            displayName={name}
-                            personalHandle={personalHandles.get(member.uid)}
-                            className="flex items-center gap-2.5 hover:opacity-80 transition-opacity"
-                          >
-                            <div className="flex items-center gap-2.5">
-                              {avatar !== undefined ? (
-                                // eslint-disable-next-line @next/next/no-img-element
-                                <img src={avatar} alt={name} className="h-7 w-7 shrink-0 rounded-full object-cover" />
-                              ) : (
-                                <div className="flex h-7 w-7 shrink-0 items-center justify-center rounded-full bg-primary/10 text-primary text-[11px] font-semibold">
-                                  {initials}
-                                </div>
-                              )}
-                              <span className="font-medium">{name}</span>
-                            </div>
-                          </UserProfileLink>
-                        </td>
-
-                        {/* Email — owner/admin only; server returns null for members */}
-                        {canManageMembers && (
-                          <td className="px-4 py-3 text-xs text-muted-foreground whitespace-nowrap">
-                            {member.email ?? '—'}
-                          </td>
-                        )}
-
-                        {/* Role */}
-                        <td className="px-4 py-3 whitespace-nowrap">
-                          {isOwner && member.role !== 'owner' ? (
-                            <button
-                              type="button"
-                              onClick={() => handleToggleRole(member.uid, member.role)}
-                              title={`Click to change to ${member.role === 'admin' ? 'member' : 'admin'}`}
-                              className="cursor-pointer"
-                            >
-                              <RoleBadge role={member.role} />
-                            </button>
-                          ) : (
-                            <RoleBadge role={member.role} />
-                          )}
-                        </td>
-
-                        {/* Joined */}
-                        <td className="px-4 py-3 text-xs text-muted-foreground whitespace-nowrap">
-                          {formatDate(member.joinedAt)}
-                        </td>
-
-                        {/* Last sign in — owner/admin only; server returns null for members */}
-                        {canManageMembers && (
-                          <td className="px-4 py-3 whitespace-nowrap text-xs text-muted-foreground">
-                            {member.lastSignInTime === null || member.lastSignInTime === undefined
-                              ? <span className="text-muted-foreground/50">Never</span>
-                              : formatLastSignIn(member.lastSignInTime)}
-                          </td>
-                        )}
-
-                        {/* Actions */}
-                        <td className="px-4 py-3 whitespace-nowrap">
-                          <div className="flex items-center justify-end gap-1">
-                            {canManageMembers && member.role !== 'owner' && member.uid !== user?.id ? (
-                              <>
-                                {/* Resend only makes sense while the invite is
-                                    still pending. Once the member has signed in
-                                    the backend refuses the resend, so hide it
-                                    rather than surface a confusing error. */}
-                                {(member.lastSignInTime === null || member.lastSignInTime === undefined) && (
-                                  <button
-                                    type="button"
-                                    onClick={() => handleResendInvite(member.uid)}
-                                    disabled={resendingUid === member.uid}
-                                    title="Resend invite"
-                                    className="rounded p-1.5 text-muted-foreground hover:text-primary hover:bg-primary/10 transition-colors disabled:opacity-40"
-                                    aria-label={`Resend invite to ${name}`}
-                                  >
-                                    <MailIcon className="h-3.5 w-3.5" />
-                                  </button>
-                                )}
-                                <button
-                                  type="button"
-                                  onClick={() => handleRemoveMember(member.uid)}
-                                  title="Remove member"
-                                  className="rounded p-1.5 text-muted-foreground hover:text-destructive hover:bg-destructive/10 transition-colors"
-                                  aria-label={`Remove ${name}`}
-                                >
-                                  <Trash2 className="h-3.5 w-3.5" />
-                                </button>
-                              </>
-                            ) : null}
-                          </div>
-                        </td>
-                      </tr>
-                    );
-                  })}
-                </tbody>
-              </table>
-            </div>
-          )}
-
-          {/* Invite form */}
-          {showInviteForm && canManageMembers && (
-            <div className="mt-4 rounded-lg border bg-card px-4 py-5">
-              <h3 className="text-sm font-semibold mb-4">Invite user</h3>
-              <form onSubmit={handleInvite} className="flex flex-col gap-3">
-                <div className="flex flex-col gap-1.5">
-                  <label htmlFor="inviteEmail" className="text-sm font-medium">
-                    Email <span className="text-destructive">*</span>
-                  </label>
-                  <input
-                    id="inviteEmail"
-                    type="email"
-                    value={inviteEmail}
-                    onChange={(e) => setInviteEmail(e.target.value)}
-                    placeholder="colleague@example.com"
-                    className="rounded-md border bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50"
-                    disabled={inviting}
-                    required
-                  />
-                </div>
-
-                <div className="flex flex-col gap-1.5">
-                  <label htmlFor="inviteName" className="text-sm font-medium">
-                    Display name
-                  </label>
-                  <input
-                    id="inviteName"
-                    type="text"
-                    value={inviteName}
-                    onChange={(e) => setInviteName(e.target.value)}
-                    placeholder="Jane Smith"
-                    className="rounded-md border bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50"
-                    disabled={inviting}
-                  />
-                </div>
-
-                <div className="flex flex-col gap-1.5">
-                  <label htmlFor="inviteRole" className="text-sm font-medium">
-                    Role
-                  </label>
-                  <select
-                    id="inviteRole"
-                    value={inviteRole}
-                    onChange={(e) => setInviteRole(e.target.value as MemberRole)}
-                    className="rounded-md border bg-background px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50"
-                    disabled={inviting}
-                  >
-                    <option value="member">member</option>
-                    <option value="admin">admin</option>
-                  </select>
-                </div>
-
-                {error !== null && (
-                  <p className="text-xs text-destructive">{error}</p>
-                )}
-
-                <div className="flex items-center gap-2 pt-1">
-                  <button
-                    type="button"
-                    onClick={() => { setShowInviteForm(false); setError(null); }}
-                    disabled={inviting}
-                    className="rounded-md border px-4 py-2 text-sm font-medium hover:bg-muted transition-colors disabled:opacity-50"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    type="submit"
-                    disabled={inviting}
-                    className="rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground hover:bg-primary/90 disabled:opacity-50 transition-colors"
-                  >
-                    {inviting ? 'Sending…' : 'Send invite'}
-                  </button>
-                </div>
-              </form>
-            </div>
-          )}
-        </div>
-
-        {/* ── Section 4: Workspace Secrets ──────────────────────────────── */}
         {canEditProfile && user?.id && (
           <div className="mb-10 space-y-4">
             <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Workspace Secrets</h2>
@@ -1136,129 +116,22 @@ export default function WorkspaceConfigPage() {
           </div>
         )}
 
-        {/* ── Section 5: Administration ──────────────────────────────────── */}
-        {canEditProfile && (
-          <div className="mb-10 space-y-4">
-            <h2 className="text-xs font-semibold uppercase tracking-wider text-muted-foreground">Administration</h2>
-            <div className="rounded-lg border bg-card px-4 py-5 space-y-3">
-              <Link
-                href={`/${handle}/admin/infrastructure`}
-                className="flex items-center justify-between group"
-              >
-                <div>
-                  <p className="text-sm font-medium group-hover:text-primary transition-colors">Infrastructure</p>
-                  <p className="text-xs text-muted-foreground">Docker images and disk usage</p>
-                </div>
-                <ArrowLeft className="h-4 w-4 text-muted-foreground rotate-180 group-hover:text-primary transition-colors" />
-              </Link>
-              <div className="border-t" />
-              <Link
-                href={routes.adminToolCatalog(handle, { from: 'settings' })}
-                className="flex items-center justify-between group"
-              >
-                <div>
-                  <p className="text-sm font-medium group-hover:text-primary transition-colors">Tool catalog</p>
-                  <p className="text-xs text-muted-foreground">MCP servers available to agents</p>
-                </div>
-                <ArrowLeft className="h-4 w-4 text-muted-foreground rotate-180 group-hover:text-primary transition-colors" />
-              </Link>
-              <div className="border-t" />
-              <Link
-                href={routes.adminOAuthProviders(handle, { from: 'settings' })}
-                className="flex items-center justify-between group"
-              >
-                <div>
-                  <p className="text-sm font-medium group-hover:text-primary transition-colors">OAuth providers</p>
-                  <p className="text-xs text-muted-foreground">External authentication for MCP tools</p>
-                </div>
-                <ArrowLeft className="h-4 w-4 text-muted-foreground rotate-180 group-hover:text-primary transition-colors" />
-              </Link>
-              <div className="border-t" />
-              <Link
-                href={`/${handle}/admin/email-status`}
-                className="flex items-center justify-between group"
-              >
-                <div>
-                  <p className="text-sm font-medium group-hover:text-primary transition-colors">Email</p>
-                  <p className="text-xs text-muted-foreground">Email provider configuration status</p>
-                </div>
-                <ArrowLeft className="h-4 w-4 text-muted-foreground rotate-180 group-hover:text-primary transition-colors" />
-              </Link>
-            </div>
-          </div>
-        )}
+        {canEditProfile && <WorkspaceAdministrationSection handle={handle} />}
 
-        {/* ── Danger zone ───────────────────────────────────────────────────── */}
         {dangerError !== null && (
           <div className="mb-3 rounded-md border border-destructive/30 bg-destructive/5 px-3 py-2 text-xs text-destructive">
             {dangerError}
           </div>
         )}
 
-        {!isOwner && currentUserMember !== undefined && (
-          <div className="rounded-lg border border-destructive/30 bg-card px-4 py-5">
-            <h2 className="text-sm font-semibold mb-1">Leave workspace</h2>
-            <p className="text-xs text-muted-foreground mb-3">
-              You will lose access to this workspace&apos;s resources.
-            </p>
-            <button
-              type="button"
-              onClick={handleLeaveWorkspace}
-              disabled={leaving}
-              className="inline-flex items-center gap-1.5 rounded-md border border-destructive/30 px-3 py-1.5 text-xs font-medium text-destructive hover:bg-destructive/10 disabled:opacity-50 transition-colors"
-            >
-              <LogOut className="h-3.5 w-3.5" />
-              {leaving ? 'Leaving…' : 'Leave workspace'}
-            </button>
-          </div>
-        )}
-
-        {canDestroyWorkspace && isPersonal === true && (
-          <div className="rounded-lg border border-destructive/30 bg-card px-4 py-5">
-            <h2 className="text-sm font-semibold mb-1 text-destructive">Reset workspace</h2>
-            <p className="text-xs text-muted-foreground mb-3">
-              This deletes every workflow in <span className="font-semibold">@{handle}</span> along with their runs and tasks, and cannot be undone. The workspace, its members and its secrets stay — a personal workspace cannot be deleted, because you always need one.
-            </p>
-            <button
-              type="button"
-              onClick={() => setDangerDialog('reset')}
-              className="inline-flex items-center gap-1.5 rounded-md border border-destructive/30 px-3 py-1.5 text-xs font-medium text-destructive hover:bg-destructive/10 transition-colors"
-            >
-              <RotateCcw className="h-3.5 w-3.5" />
-              Reset workspace
-            </button>
-          </div>
-        )}
-
-        {isOwner && isPersonal === false && (
-          <div className="rounded-lg border border-destructive/30 bg-card px-4 py-5">
-            <h2 className="text-sm font-semibold mb-1 text-destructive">Delete workspace</h2>
-            <p className="text-xs text-muted-foreground mb-3">
-              This will permanently delete <span className="font-semibold">@{handle}</span>, remove all members, destroy every workflow, run, task and secret it holds, and cannot be undone.
-            </p>
-            <button
-              type="button"
-              onClick={() => setDangerDialog('delete')}
-              className="inline-flex items-center gap-1.5 rounded-md border border-destructive/30 px-3 py-1.5 text-xs font-medium text-destructive hover:bg-destructive/10 transition-colors"
-            >
-              <Trash2 className="h-3.5 w-3.5" />
-              Delete workspace
-            </button>
-          </div>
-        )}
-
-        {dangerDialog !== null && (
-          <WorkspaceDangerDialog
-            mode={dangerDialog}
-            handle={handle}
-            open
-            onOpenChange={(value) => { if (!value) setDangerDialog(null); }}
-            onDone={() => {
-              setDangerDialog(null);
-              if (dangerDialog === 'delete') router.push('/workspace-selection');
-            }}
-          />
-        )}
+        <WorkspaceDangerZone
+          handle={handle}
+          isMember={currentUserMember !== undefined}
+          isOwner={isOwner}
+          isPersonal={isPersonal}
+          canDestroyWorkspace={canDestroyWorkspace}
+          onError={setDangerError}
+        />
       </div>
     </div>
   );
